ruby: manually manage packet_buffer lifetime
The trilogy client eagerly allocate a 32kiB buffer, and grows
it as needed. It's never freed not shrunk until the connection
is closed.

For basic usage it's not a big deal, but some applications may
have dozens if not hundreds of connections that are mostly idle.
A common case being multi-tenant applications with horizontal
sharding. In such cases you only ever query one database but
have open connections to many databases.

This situation might lead to a lot of memory retained by trilogy
connections and never really released.

Ideally you'd want to use a buffer pool, but because we call
`rb_ext_ractor_safe(true)`, we'd need such pool to be ractor
safe, which means either being synchronized or having one pool
per ractor.

This commit simply malloc and free the buffer directly to prove
the approach works, and a pool could be fitted on top.

Before:

```ruby
>> t = Trilogy.new(database: "test")
>> ObjectSpace.memsize_of(t)
=> 66176
>> t.query("select '#{"a" * 10_000_000}' as a"); nil
>> ObjectSpace.memsize_of(t)
=> 16810624
```

After:

```ruby
>> t = Trilogy.new(database: "test")
>> ObjectSpace.memsize_of(t)
=> 33408
>>  t.query("select '#{"a" * 10_000_000}' as a"); nil
>> ObjectSpace.memsize_of(t)
=> 33408
```